Refactor CLI into agent backends
