+++ src/coding_review_agent_loop/agents/__init__.py
@@ -1,0 +1,1 @@
+"""Agent backend implementations."""
