+++ src/coding_review_agent_loop/errors.py
@@ -1,0 +1,7 @@
+"""Shared exceptions for the agent loop."""
+
+from __future__ import annotations
+
+
+class AgentLoopError(RuntimeError):
+    """Raised for expected orchestration failures."""
